change type: int->double

+++ Project3/main.cpp
@@ -2,7 +2,7 @@
 
 class Cal {
 public:
-	int getSum(int a, int b) {
+	double getSum(double a, double b) {
 		return a + b;
 	}
 };
